MRM 02_03 Creating an Express Server
A Demonstration of how express servers can be used to send strings of data into the client requesting the data from a specific routing path.

This application of sending strings of data is applicable to API calls when a client request for data to a specific routing path that will equate to data calling from database.